Rename to permissionsManager
