+++ backend/src/API/candidateDeciderAPI.ts
@@ -1,6 +1,6 @@
 import CandidateDeciderDao from '../dao/CandidateDeciderDao';
 import { NotFoundError, PermissionError } from '../utils/errors';
-import PermissionsManager from '../utils/permissions';
+import PermissionsManager from '../utils/permissionsManager';
 
 export const getAllCandidateDeciderInstances = async (
   user: IdolMember
