+++ backend/src/siteIntegration.ts
@@ -1,6 +1,6 @@
 import { Octokit } from '@octokit/rest';
 import { PRResponse } from './GithubTypes';
-import PermissionsManager from './utils/permissions';
+import PermissionsManager from './utils/permissionsManager';
 import { PermissionError, BadRequestError } from './utils/errors';
 
 require('dotenv').config();
