+++ backend/src/utils/permissionsManager.ts
